+++ src/components/Cart/Cart.js
@@ -1,22 +1,17 @@
-import React, { Component } from 'react';
-import './Cart.css';
 import { currentDiscount, currentDiscountedElements } from '../Main/Main';
 import { catalog } from '../../utilities/catalog';
 import CartButtons from '../Buttons/CartButtons/CartButtons';
 import Cards from '../Card/Cards';
 import Button from '../UI/Button/Button';
+import './Cart.css';
 
 
-class Cart extends Component {
+const Cart = props => {
 
-  componentDidMount() {
-
-  }
-
-  goToCheckout = () => {
+  const goToCheckout = () => {
     const authors = {}
 
-    let totalDue = this.props.cart.reduce((acc, curr) => {
+    let totalDue = props.cart.reduce((acc, curr) => {
       let target = catalog.find(el => el.title === curr);
       authors[target.title] = target.author;
       let targetPrice = target.price;
@@ -24,7 +19,7 @@ class Cart extends Component {
       return acc + targetPrice
     }, 0)
 
-    let items = [...this.props.cart].reduce((accum, book) => {
+    let items = [...props.cart].reduce((accum, book) => {
       if (!accum[book]) {
         accum[book] = 1;
         return accum
@@ -39,46 +34,44 @@ class Cart extends Component {
     }
     books.push('totalPrice=' + encodeURIComponent(totalDue.toFixed(2)))
 
-    this.props.history.push({
+    props.history.push({
       pathname: '/checkout',
       search: '?' + books.join('&')
     })
   }
 
-  render() {
-    let totalDue = this.props.cart.reduce((acc, curr) => {
-      let target = catalog.find(el => el.title === curr);
-      let targetPrice = target.price;
-      if (currentDiscountedElements.includes(target.title)) { targetPrice = Math.round(targetPrice - ((targetPrice / 100) * currentDiscount)) }
-      return acc + targetPrice
-    }, 0)
+  let totalDue = props.cart.reduce((acc, curr) => {
+    let target = catalog.find(el => el.title === curr);
+    let targetPrice = target.price;
+    if (currentDiscountedElements.includes(target.title)) { targetPrice = Math.round(targetPrice - ((targetPrice / 100) * currentDiscount)) }
+    return acc + targetPrice
+  }, 0)
 
-    let productsAndTotal = <div style={{ width: '100%', height: '100%' }}>
-      <Cards
-        cart={this.props.cart}
-        showDescription={this.props.showDescription}
-        adjustQuantity={this.props.adjustQuantity}
-      />
-      <p id='totalDue'>
-        <span>Total: </span>{`${totalDue.toFixed(2)}`}$
-        <Button click={this.goToCheckout} btnClass={['checkout-btn']} >
-          Checkout
+  let productsAndTotal = <div style={{ width: '100%', height: '100%' }}>
+    <Cards
+      cart={props.cart}
+      showDescription={props.showDescription}
+      adjustQuantity={props.adjustQuantity}
+    />
+    <p id='totalDue'>
+      <span>Total: </span>{`${totalDue.toFixed(2)}`}$
+        <Button click={goToCheckout} btnClass={['checkout-btn']} >
+        Checkout
           </Button>
-      </p>
-    </div>
+    </p>
+  </div>
 
-    return (
-      <div id='cart-box-wrapper' className={this.props.fullPage ? 'full-view' : 'normal-view'} >
-        <CartButtons
-          clickFull={this.props.fullView}
-          clickClose={this.props.clickToggle}
-        />
-        <div id='inner-cart-box'>
-          {totalDue ? productsAndTotal : <p id='empty-cart-message'>You have currently no items in the cart</p>}
-        </div>
+  return (
+    <div id='cart-box-wrapper' className={props.fullPage ? 'full-view' : 'normal-view'} >
+      <CartButtons
+        clickFull={props.fullView}
+        clickClose={props.clickToggle}
+      />
+      <div id='inner-cart-box'>
+        {totalDue ? productsAndTotal : <p id='empty-cart-message'>You have currently no items in the cart</p>}
       </div>
-    )
-  }
+    </div>
+  )
 
 }
